new trait for KB Matrix
