+++ src/main/scala/cc/factorie/epistemodb/tac/TrainTestTacData.scala
@@ -67,7 +67,7 @@ object TrainTestTacData {
       opts.parse(args)
 
       val tReadStart = System.currentTimeMillis
-      val kb = KBMatrix.fromTsv(opts.tacData.value).prune(2,1)
+      val kb = EntityRelationKBMatrix.fromTsv(opts.tacData.value).prune(2,1)
       val tRead = (System.currentTimeMillis - tReadStart)/1000.0
       println(f"Reading from file and pruning took $tRead%.2f s")
