+++ src/test/scala/cc/factorie/epistemodb/TestEntityRelationKBMatrix.scala
@@ -11,12 +11,12 @@ import scala.util.Random
 /**
  * Created by beroth on 2/6/15.
  */
-class TestKBMatrix extends JUnitSuite  with util.FastLogging  {
+class TestEntityRelationKBMatrix extends JUnitSuite  with util.FastLogging  {
 
   val eps = 1e-4
 
   @Test def getSetCellsTest() {
-    val m = new KBMatrix()
+    val m = new EntityRelationKBMatrix()
     m.set("Barack Obama", "Michelle Obama", "is married to", 5.0)
     m.set("Barack Obama", "Michelle Obama", "is married to", 10.0)
     m.set("Barack Obama", "Michelle Obama", "per:spouse", 1.0)
@@ -33,7 +33,7 @@ class TestKBMatrix extends JUnitSuite  with util.FastLogging  {
   }
 
   @Test def equalsTest() {
-    val m1 = new KBMatrix()
+    val m1 = new EntityRelationKBMatrix()
     m1.set("Barack Obama", "Michelle Obama", "is married to", 5.0)
     m1.set("Barack Obama", "Michelle Obama", "is married to", 10.0)
     m1.set("Barack Obama", "Michelle Obama", "per:spouse", 1.0)
@@ -41,21 +41,21 @@ class TestKBMatrix extends JUnitSuite  with util.FastLogging  {
     m1.set("Frank Sinatra", "Nancy Barbato", "and his wife", 2.0)
 
     // same as m1, but constructed in different order
-    val m2 = new KBMatrix()
+    val m2 = new EntityRelationKBMatrix()
     m2.set("Frank Sinatra", "Nancy Barbato", "and his wife", 2.0)
     m2.set("Frank Sinatra", "Nancy Barbato", "per:spouse", 1.0)
     m2.set("Barack Obama", "Michelle Obama", "per:spouse", 1.0)
     m2.set("Barack Obama", "Michelle Obama", "is married to", 10.0)
 
     // similar to m2, but one different cell value
-    val m3 = new KBMatrix()
+    val m3 = new EntityRelationKBMatrix()
     m3.set("Frank Sinatra", "Nancy Barbato", "and his wife", 1.0)
     m3.set("Frank Sinatra", "Nancy Barbato", "per:spouse", 1.0)
     m3.set("Barack Obama", "Michelle Obama", "per:spouse", 1.0)
     m3.set("Barack Obama", "Michelle Obama", "is married to", 10.0)
 
     // different rows/columns
-    val m4 = new KBMatrix()
+    val m4 = new EntityRelationKBMatrix()
     m4.set("Frank Sinatra", "Nancy Barbato", "and his wife", 2.0)
     m4.set("Frank Sinatra", "Nancy Barbato", "per:spouse", 1.0)
     m4.set("Nicola Sarcozy", "Carla Bruni", "per:spouse", 1.0)
@@ -72,7 +72,7 @@ class TestKBMatrix extends JUnitSuite  with util.FastLogging  {
     val fongo = new Fongo("myserver");
     val db : DB = fongo.getDB("mydb");
 
-    val m1 = new KBMatrix()
+    val m1 = new EntityRelationKBMatrix()
     m1.set("Frank Sinatra", "Nancy Barbato", "and his wife", 2.0)
     m1.set("Frank Sinatra", "Nancy Barbato", "per:spouse", 1.0)
     m1.set("Barack Obama", "Michelle Obama", "per:spouse", 1.0)
@@ -80,7 +80,7 @@ class TestKBMatrix extends JUnitSuite  with util.FastLogging  {
 
     m1.writeToMongo(db)
 
-    val m2 = KBMatrix.fromMongo(db)
+    val m2 = EntityRelationKBMatrix.fromMongo(db)
     assertTrue(m1.hasSameContent(m2))
   }
 
@@ -90,7 +90,7 @@ class TestKBMatrix extends JUnitSuite  with util.FastLogging  {
     val fongo = new Fongo("myserver");
     val db : DB = fongo.getDB("mydb");
 
-    val m1 = new KBMatrix()
+    val m1 = new EntityRelationKBMatrix()
     m1.set("Frank Sinatra", "Nancy Barbato", "and his wife", 2.0)
     m1.set("Frank Sinatra", "Nancy Barbato", "per:spouse", 1.0)
     m1.set("Barack Obama", "Michelle Obama", "per:spouse", 1.0)
@@ -98,12 +98,12 @@ class TestKBMatrix extends JUnitSuite  with util.FastLogging  {
 
     m1.writeToMongoCellBased(db)
 
-    val m2 = KBMatrix.fromMongoCellBased(db)
+    val m2 = EntityRelationKBMatrix.fromMongoCellBased(db)
     assertTrue(m1.hasSameContent(m2))
   }
 
   @Test def pruneMatrixTest() {
-    val m = new KBMatrix()
+    val m = new EntityRelationKBMatrix()
     m.set("Barack Obama", "Michelle Obama", "is married to", 1.0)
     m.set("Frank Sinatra", "Nancy Barbato", "per:spouse", 1.0)
     m.set("Frank Sinatra", "Nancy Barbato", "and his wife", 1.0)
@@ -111,7 +111,7 @@ class TestKBMatrix extends JUnitSuite  with util.FastLogging  {
 
     val m0 = m.prune(0,0)
 
-    val m0goal = new KBMatrix()
+    val m0goal = new EntityRelationKBMatrix()
     m0goal.set("Frank Sinatra", "Nancy Barbato", "per:spouse", 1.0)
     m0goal.set("Frank Sinatra", "Nancy Barbato", "and his wife", 1.0)
     m0goal.set("Nicola Sarcozy", "Carla Bruni", "and his wife", 1.0)
@@ -119,14 +119,14 @@ class TestKBMatrix extends JUnitSuite  with util.FastLogging  {
     assertTrue(m0.hasSameContent(m0goal))
 
     val m1 = m.prune(0,1)
-    val m1goal = new KBMatrix()
+    val m1goal = new EntityRelationKBMatrix()
     m1goal.set("Frank Sinatra", "Nancy Barbato", "and his wife", 1.0)
     m1goal.set("Nicola Sarcozy", "Carla Bruni", "and his wife", 1.0)
 
     assertTrue(m1.hasSameContent(m1goal))
 
     val m2 = m.prune(1,0)
-    val m2goal = new KBMatrix()
+    val m2goal = new EntityRelationKBMatrix()
     m2goal.set("Frank Sinatra", "Nancy Barbato", "per:spouse", 1.0)
     m2goal.set("Frank Sinatra", "Nancy Barbato", "and his wife", 1.0)
     assertTrue(m2.hasSameContent(m2goal))
@@ -137,7 +137,7 @@ class TestKBMatrix extends JUnitSuite  with util.FastLogging  {
     //1101
     //0010
     //1101
-    val m = new KBMatrix()
+    val m = new EntityRelationKBMatrix()
     m.set("A", "A", "1",1.0)
     m.set("A", "A", "3",1.0)
